chore: update texts and posts order on variable page
