+++ src/app/variable-load/page.jsx
@@ -13,104 +13,9 @@ import Layout from 'components/shared/layout';
 import TableOfContents from 'components/shared/table-of-contents';
 import LINKS from 'constants/links';
 import SEO_DATA from 'constants/seo-data';
-import evolutionOfPostgres from 'images/pages/variable-load/relevant-articles/evolution-of-postgres.jpg';
-import { getWpPostBySlug } from 'utils/api-posts';
-import getFormattedDate from 'utils/get-formatted-date';
 import getMetadata from 'utils/get-metadata';
 
 export const metadata = getMetadata(SEO_DATA.variable);
-
-const blogArticlesData = [
-  {
-    slug: '/blog/autoscaling-in-action-postgres-load-testing-with-pgbench',
-    order: 2,
-  },
-  {
-    slug: '/blog/1-year-of-autoscaling-postgres-at-neon',
-    order: 3,
-  },
-  {
-    slug: '/blog/white-widgets-secret-to-scalable-postgres-neon',
-    order: 9,
-  },
-  {
-    slug: '/blog/how-recrowd-uses-neon-autoscaling-to-meet-fluctuating-demand',
-    order: 10,
-  },
-  {
-    slug: '/blog/why-you-want-a-database-that-scales-to-zero',
-    order: 11,
-  },
-];
-
-const docsArticlesData = [
-  {
-    slug: '/docs/introduction/autoscaling',
-    title: 'Autoscaling',
-    order: 6,
-  },
-  {
-    slug: '/docs/extensions/neon-utils',
-    title: 'The neon_utils extension',
-    order: 7,
-  },
-];
-
-const externalArticlesData = [
-  {
-    url: 'https://www.outerbase.com/blog/the-evolution-of-serverless-postgres/',
-    title: 'The Evolution of Serverless Postgres',
-    image: evolutionOfPostgres,
-    date: 'May 29, 2024',
-    order: 1,
-  },
-  {
-    url: 'https://medium.com/@carlotasotos/database-economics-an-amazon-rds-reflection-5d7a35638b20',
-    title: 'Database economics: an Amazon RDS reflection',
-    date: 'May 31, 2024',
-    order: 4,
-  },
-  {
-    url: 'https://dev.to/rsiv/auto-scaling-apps-by-default-neon-api-gateway-nitric-34id',
-    title: 'Architectural choices that scale',
-    date: 'Jun 6, 2023',
-    order: 5,
-  },
-  {
-    url: 'https://github.com/prisma/read-replicas-demo',
-    title: 'Read Replicas Demo',
-    order: 8,
-  },
-];
-
-const fetchBlogArticles = async (articles) =>
-  Promise.all(
-    articles.map(async (article) => {
-      const { post } = await getWpPostBySlug(article.slug);
-      if (post) {
-        return {
-          ...article,
-          title: post.title,
-          date: getFormattedDate(post.date),
-          image: post.seo?.twitterImage?.mediaItemUrl,
-        };
-      }
-      return article;
-    })
-  );
-
-const articlesWithImages = (articles) =>
-  articles.map((article) => {
-    if (article.image) {
-      return article;
-    }
-    const encodedTitle = Buffer.from(article.title).toString('base64');
-    const imagePath = `/docs/og?title=${encodedTitle}&show-logo=${article.url ? 'false' : 'true'}`;
-    return {
-      ...article,
-      image: imagePath,
-    };
-  });
 
 const titles = [
   'Variable traffic, fixed costs?',
@@ -120,11 +25,6 @@ const titles = [
 ];
 
 const VariableLoadPage = async () => {
-  const blogArticles = await fetchBlogArticles(blogArticlesData);
-  const docsArticles = articlesWithImages(docsArticlesData);
-  const externalArticles = articlesWithImages(externalArticlesData);
-  const allArticles = [...blogArticles, ...docsArticles, ...externalArticles];
-
   const tableOfContents = titles.map((title) => ({
     title,
     id: slugify(title.replace(/&nbsp;/g, ' '), {
@@ -150,7 +50,6 @@ const VariableLoadPage = async () => {
               <Unique title={tableOfContents[3]} />
             </article>
           </div>
-
           <div className="col-start-10 col-end-13 ml-[50px] h-full xl:hidden">
             <nav
               className={clsx(
@@ -162,8 +61,7 @@ const VariableLoadPage = async () => {
             </nav>
           </div>
         </Container>
-
-        <RelevantArticles articles={allArticles} />
+        <RelevantArticles />
         <Cta
           className="pb-[290px] pt-[285px] xl:py-[230px] lg:pb-[156px] lg:pt-[179px] sm:pb-[110px] sm:pt-[116px]"
           title="Try it yourself"
