+++ src/components/pages/variable/unique/unique.jsx
@@ -9,34 +9,30 @@ import Testimonial from '../testimonial';
 
 const items = [
   {
-    text: 'Neon compute costs are <a href="https://www.outerbase.com/blog/the-evolution-of-serverless-postgres/" target="_blank" rel="noopener noreferrer">up to 75% cheaper</a> vs Aurora Serverless v2.',
+    text: 'Neon compute costs are <a href="https://www.outerbase.com/blog/the-evolution-of-serverless-postgres/" target="_blank" rel="noopener noreferrer">up to 75% cheaper</a> vs Aurora&nbsp;Serverless&nbsp;v2.',
   },
   {
-    text: 'Neon <strong>scales to zero</strong>, Aurora Serverless does not.',
+    text: 'Neon scales to zero, Aurora Serverless does not.',
   },
   {
-    text: 'Neon <strong>provisions instances in < 1 s</strong>, compared to Aurora&apos;s up to&nbsp;20&nbsp;min.',
+    text: 'Neon provisions instances in < 1 s, compared to Aurora&apos;s up to&nbsp;20&nbsp;min.',
   },
   {
-    text: 'Neon uses <strong>transparent compute units</strong>, vs the ACU abstraction in&nbsp;Aurora&nbsp;Serverless.',
+    text: 'Neon uses transparent compute units, vs the ACU abstraction in&nbsp;Aurora&nbsp;Serverless.',
   },
   {
-    text: 'Neon supports <strong>database branching with data and schema via copy-on-write,</strong> <a href="/flow">improving development workflows.</a>',
+    text: 'Neon supports database branching with data and schema via copy-on-write, <a href="/flow">improving development workflows.</a>',
   },
   {
-    text: 'Neon&apos;s <strong>read replicas don&apos;t require storage redundancy</strong>, differently than Aurora&apos;s.',
+    text: 'Neon&apos;s read replicas don&apos;t require storage redundancy, differently than Aurora&apos;s.',
   },
   {
-    text: '<strong>Connection pooling is built-in in Neon,</strong> vs Aurora&apos;s RDS Proxy.',
+    text: 'Connection pooling is built-in in Neon, vs Aurora&apos;s RDS Proxy.',
   },
 ];
 
 const Unique = ({ title }) => (
   <Section className="unique" title={title}>
-    <div className="prose-variable">
-      <p>The Neon architecture is inspired in Amazon&nbsp;Aurora, but with some key differences:</p>
-      <List items={items} />
-    </div>
     <Testimonial
       text="Before choosing Neon, we also considered Aurora, but the opacity of&nbsp;the pricing model did not convince us and costs seemed to&nbsp;rise&nbsp;quickly."
       author={{
@@ -46,6 +42,10 @@ const Unique = ({ title }) => (
       }}
       url={`${LINKS.blog}/white-widgets-secret-to-scalable-postgres-neon`}
     />
+    <div className="prose-variable">
+      <p>The Neon architecture is inspired in Amazon&nbsp;Aurora, but with some key differences:</p>
+      <List items={items} />
+    </div>
   </Section>
 );
